Use full member objects, if available
- Moves TYPING_START to an Action to be able to use this.getMember()
